instruction-level generic join
also move some stuff out of simple_value
